move utilities to shared module
