+++ src/app/pages/docker-run/docker-run.component.ts
@@ -2,8 +2,8 @@ import { Component, ElementRef, OnInit, ViewChild } from '@angular/core';
 import { UntypedFormArray, UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { LoggingService } from 'src/app/modules/shared/services/logging.service';
-import { StringBuilder } from 'src/app/utilities/string-builder';
 import { FormService } from '../../modules/shared/services/form-service.service';
+import { StringBuilder } from '../../modules/shared/utilities/string-builder';
 import { requiredIfValidator } from '../../modules/shared/validators/required-if.validator';
 
 export class DockerRunModel {
